# openapi.yaml
openapi: 3.0.3
info:
  title: 災害物資需求 API
  version: v1.1.0
  description: |-
    依據需求圖片實作的後端 API。提供建立物資需求、查詢需求清單、物資配送登記。
servers:
  - url: http://localhost:8080
    description: 本地開發
  - url: https://guangfu250923.pttapp.cc
    description: 線上服務 (Production)
  - url: '{protocol}://{host}:{port}'
    description: 自訂伺服器
    variables:
      protocol:
        default: http
        enum: [http, https]
      host:
        default: localhost
      port:
        default: '8080'
paths:
...
  /supplies:
    get:
      operationId: listSupplies
      summary: 取得供應單清單 (分頁)
      description: 列出所有 supplies 供應單。
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: embed
          schema:
            type: string
            enum: [all]
          description: 若設為 all，回傳集合中每個供應單的 supplies 會嵌入其全部物資項目；未指定時 supplies 為空陣列（僅佔位），需再以 GET /supplies/{id} 取得詳細。
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/SupplyCollection' } } } }
    post:
      operationId: createSupply
      summary: 建立供應單
      description: 建立一筆新的供應單；可同時附上一個第一筆物資項目 (supplies)。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SupplyCreate' }
      responses:
        '201': { description: 建立成功, content: { application/json: { schema: { $ref: '#/components/schemas/Supply' } } } }
        '400': { description: 輸入錯誤 }
  /supplies/{id}:
    get:
      operationId: getSupply
      summary: 取得單一供應單
      description: 依供應單 UUID 取得完整供應單資訊，並回傳其所有物資項目 (supplies 陣列，可能為空)。
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/Supply' } } } }
        '404': { description: 找不到 }
    patch:
      operationId: patchSupply
      summary: 更新供應單 (部分欄位) (停用)
      description: 對供應單進行部分欄位更新；僅更新傳入的欄位 (name/address/phone/notes)。不影響其下物資項目。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SupplyPatch' }
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: 更新成功, content: { application/json: { schema: { $ref: '#/components/schemas/Supply' } } } }
        '400': { description: 輸入錯誤 }
        '404': { description: 找不到 }
    post:
      operationId: distributeSupplyItems
      summary: 批次配送 (累加 recieved_count)
      description: 對指定供應單底下的多個物資項目增加配送數量 (更新 recieved_count)。避免超過 total_count。
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required: [id,count]
                properties:
                  id: { type: string, format: uuid, description: supply_item ID }
                  count: { type: integer, minimum: 1, description: 本次配送新增的數量 }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/SupplyItem' } } } } }
        '400': { description: 輸入錯誤或超過需求 }
        '404': { description: 找不到 }
  /supply_items:
    get:
      operationId: listSupplyItems
      summary: 取得物資項目清單 (分頁)
      description: 分頁列出所有物資項目，可用 supply_id 過濾特定供應單；採 JSON-LD Collection 格式。
      parameters:
        - in: query
          name: supply_id
          schema: { type: string, format: uuid }
          description: 過濾指定供應單底下的項目
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/SupplyItemCollection' } } } }
    post:
      operationId: createSupplyItem
      summary: 建立物資項目
      description: 為既有供應單新增一筆物資項目；初始 recieved_count 預設為 0。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SupplyItemCreate' }
      responses:
        '201': { description: 建立成功, content: { application/json: { schema: { type: object, properties: { id: { type: string, format: uuid } } } } } }
        '400': { description: 輸入錯誤 }
  /supply_items/{id}:
    get:
      operationId: getSupplyItem
      summary: 取得單一物資項目
      description: 依物資項目 UUID 取得其詳細資訊。
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/SupplyItem' } } } }
        '404': { description: 找不到 }
    patch:
      operationId: patchSupplyItem
      summary: 更新物資項目 (部分欄位) (停用)
      description: 部分更新物資項目欄位；若更新 recieved_count 則不得超過 total_count。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SupplyItemPatch' }
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: 更新成功, content: { application/json: { schema: { $ref: '#/components/schemas/SupplyItem' } } } }
        '400': { description: 輸入錯誤 }
        '404': { description: 找不到 }
components:
  schemas:
    CollectionBase:
      type: object
      properties:
        '@context': { type: string, example: https://www.w3.org/ns/hydra/context.jsonld }
        '@type': { type: string, example: Collection }
        totalItems: { type: integer, description: 符合條件的總筆數 }
        limit: { type: integer, description: 本次回傳的筆數上限 }
        offset: { type: integer, description: 本次回傳的起始位置 }
        next: { type: string, nullable: true, description: 下一頁的連結 (若有) }
        previous: { type: string, nullable: true, description: 前一頁的連結 (若有) }
...
    Supply:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, nullable: true }
        address: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        pii_date: { type: integer, format: int64, nullable: true, description: 個資同意時間 (Unix Timestamp) }
        created_at: { type: integer, format: int64 }
        updated_at: { type: integer, format: int64 }
        supplies:
          type: array
          description: 供應單全部物資項目 (可能為空陣列)
          items: { $ref: '#/components/schemas/SupplyItem' }
    SupplyCreate:
      type: object
      properties:
        name: { type: string, nullable: true }
        address: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        pii_date: { type: integer, format: int64, nullable: true, description: 個資同意時間 (Unix Timestamp) }
        supplies:
          type: object
          nullable: true
          description: 同時建立的第一個物資項目
          properties:
            tag: { type: string, nullable: true }
            name: { type: string, nullable: true }
            recieved_count: { type: integer, description: '前端拼字 (received_count)；可省略預設 0' }
            total_count: { type: integer }
            unit: { type: string, nullable: true }
    SupplyPatch:
      type: object
      properties:
        name: { type: string, nullable: true }
        address: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        pii_date: { type: integer, format: int64, nullable: true, description: 個資同意時間 (Unix Timestamp) }
    SupplyCollection:
      allOf:
        - $ref: '#/components/schemas/CollectionBase'
        - type: object
          properties:
            member:
              type: array
              items: { $ref: '#/components/schemas/Supply' }
    SupplyItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        supply_id: { type: string, format: uuid }
        tag: { type: string, nullable: true }
        name: { type: string, nullable: true }
        recieved_count: { type: integer }
        total_count: { type: integer }
        unit: { type: string, nullable: true }
    SupplyItemCreate:
      type: object
      required: [supply_id,total_count]
      properties:
        supply_id: { type: string, format: uuid }
        tag: { type: string, nullable: true }
        name: { type: string, nullable: true }
        total_count: { type: integer }
        unit: { type: string, nullable: true }
    SupplyItemPatch:
      type: object
      properties:
        tag: { type: string, nullable: true }
        name: { type: string, nullable: true }
        recieved_count: { type: integer }
        total_count: { type: integer }
        unit: { type: string, nullable: true }
    SupplyItemCollection:
      allOf:
        - $ref: '#/components/schemas/CollectionBase'
        - type: object
          properties:
            member:
              type: array
              items: { $ref: '#/components/schemas/SupplyItem' }
...